Replace custom cache logic with WANObjectCache (WLDR-269)
This simplifies some of the logic. Additionally this adds
process cache support. Profiling indicated that a signficant chunk
of time was spent fetching the same item from cache over and over
again (7%). Additionally this removes the logic that was re-setting the
cache key on every fetch. Profiling indicated that this was taking
about 12% of the time.

The WANObjectCache refresh logic is a little smarter, where if
a cache entry is almost expired, it will refresh it after the
page has been sent, so that it is non-obvious to the user.

This also changes the extension.json defaults to be the same
as production use of this extension.

Removes the NCBITaxonomyLookupCacheRandomizeTTL config, since
WANObjectCache should do that automatically.

Change-Id: Ic0d1feed43ae17387fed3a7596ce7e0f27b1750e