Prefer using curl and try to reuse the connection where possible (WLDR-271)
If you keep the curl handle around, and the connection supports
keep-alive, you should be able to reuse the connection, which
should save time from TCP/TLS handshake.

Tested locally with an example containing 7 unique taxonomy lookups.
It seemed to save about 0.8 seconds, but hard to tell as response
times were quite variable. I checked in wireshark, and it did seem
to be reusing the connection. The real effect would probably be
more visible on pages with much larger number of lookups.

Also set a user-agent (general best practise). Add some code
to give 1 retry if recieving a 500 (In my test i noticed an
occasionally transient 500 from this api). Set a connect timeout
for curl in addition to the normal timeout.

Change-Id: I6901ab1621894335a2f706939d24c13202b7e31f